requirements.txt
appnope==0.1.4
asgiref==3.9.1
asttokens==3.0.0
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.1.8
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
distlib==0.4.0
Django==5.2.6
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.19.1
fonttools==4.56.0
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
matplotlib==3.10.0
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
nltk==3.9.1
numby==0.2.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
pyzmq==26.2.1
regex==2024.11.6
requests==2.32.5
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.1
urllib3==2.5.0
virtualenv==20.34.0
wcwidth==0.2.13
django-allauth>=0.63.0
PyJWT>=2.9.0
cryptography>=43.0.0
dj-database-url
gunicorn
uvicorn[standard]
whitenoise
psycopg2-binary